Complete security and code quality audit fixes
Backend fixes:
- Fix missing flush() in admin_router create_tenant and toggle_company_status
- Add UUID validation error handling in tenant middleware
- Add password strength validation (uppercase, lowercase, digit) in auth schemas
- Add max_length constraint to LoginRequest password field
- Move inline imports to top of file (auth/service.py, auth/router.py, auth/admin_router.py)
- Add audit logging for impersonation and company status changes
- Apply security headers to ALL responses including /api/ routes
- Fix HTTPS redirect to use proper RedirectResponse with 307 status
- Add IntegrityError handling in database.py to prevent info leakage
- Fix deprecated datetime.utcnow() calls in dashboard/router.py
- Fix float→Decimal type hints in load.py and driver.py models
- Add idempotency to settlement payment (accounting/service.py)
- Fix sensitive error message in reportlab import failure
- Fix pytest.ini format (was using invalid TOML syntax in INI file)
- Fix list() method name shadowing builtin in 5 repository files

Frontend fixes:
- Fix auth token refresh race condition with singleton promise pattern
- Add security headers (X-Content-Type-Options, X-Frame-Options, etc.) in next.config.ts
- Use generic error message for login failures (prevent user enumeration)
- Add global error boundary (app/error.tsx)

Agent-Logs-Url: https://github.com/Khomurod/The-TMS-platform/sessions/ee41f94e-f967-44eb-8f2d-bf0182c97270

+++ backend/app/models/driver.py
@@ -3,6 +3,7 @@
 import enum
 import uuid
 from datetime import date, datetime
+from decimal import Decimal
 
 from sqlalchemy import Boolean, Date, DateTime, Enum, Integer, Numeric, String, func
 from sqlalchemy.dialects.postgresql import UUID
@@ -72,7 +73,7 @@ class Driver(Base, TenantMixin):
         Enum(PayRateType, name="pay_rate_type_enum", create_constraint=True),
         nullable=True,
     )
-    pay_rate_value: Mapped[float | None] = mapped_column(
+    pay_rate_value: Mapped[Decimal | None] = mapped_column(
         Numeric(10, 4), nullable=True  # e.g., 0.65 CPM or 80%
     )
     use_company_defaults: Mapped[bool] = mapped_column(
